PR Feedback: don't publicize the FileManager extension
